backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
openai
python-dotenv
chromadb
sentence-transformers
scikit-learn